requirements.txt
Django==4.1.5
asgiref==3.6.0
sqlparse==0.4.3
psycopg-binary==3.1.8
psycopg2==2.9.5

pytz==2022.7.1
djangorestframework==3.14.0
django_celery_beat==2.4.0
pip==22.0.2
setuptools==59.6.0
celery[redis]==5.2.7
flower==1.2.0
requests==2.28.2